fix(视频播放页): 修复视频流信息显示信息不准确的问题，AVPlayer内核下播放器缓冲速度改为使用HLS Bridge的实际缓冲速度而不是AVPlayer获取到的本地loopback载入速度

+++ PiliPod/Services/OfflineCacheService.swift
@@ -228,7 +228,7 @@ private struct OfflineCacheManifest: Codable {
     let audioFileName: String
     let width: Int
     let height: Int
-    let fps: Int
+    let fps: Double
     let videoBitrate: Int
     let audioBitrate: Int
     let videoCodec: String
